requirements.txt
fastapi
uvicorn
requests
maxapi
gspread
google-auth
